feat: manifest v3 + base support for chat settings
Extension migrated to the devil Manifest V3. Tested on last version of
Edge and Firefox ; should work with the Chromium-world.
Works on Waterfox, dunno about Pale Moon or LibreWolf.

For accessibility purposes, the font size setting is now respected.
Timestamp toggling is also honored.

Note: feature is not reactive. Settings updated => a refresh is needed.
Also, I tried to implement a service worker for an isolated observer,
but since we do not really need it, I threw it.
Manifest v2/v3 incompatibilies
are boring to handle.